Added Odometry
bugfixes

+++ src/main/java/frc/robot/subsystems/ControlPanelSubsystem.java
@@ -169,7 +169,7 @@ public class ControlPanelSubsystem extends SubsystemBase {
     motor.set(speed.getAsDouble());
   }
 
-  public int getEncoderValue() {
+  public double getEncoderValue() {
     return motor.getSelectedSensorPosition();
   }
